pyproject.toml
[project]
name = "youtube-video-sentiment-analyser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.185.0",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "streamlit>=1.50.0",
    "textblob>=0.19.0",
    "torch>=2.9.0",
    "transformers>=4.57.1",
]
